// frontend/src/models/insightsModelType.ts
export interface insightsModelType extends Logic {
    actionCreators: {
        renameInsight: (item: QueryBasedInsightModel) => {
            type: 'rename insight (models.insightsModel)'
            payload: {
                item: QueryBasedInsightModel
            }
        }
        renameInsightSuccess: (item: QueryBasedInsightModel) => {
            type: 'rename insight success (models.insightsModel)'
            payload: {
                item: QueryBasedInsightModel
            }
        }
        duplicateInsight: (item: QueryBasedInsightModel) => {
            type: 'duplicate insight (models.insightsModel)'
            payload: {
                item: QueryBasedInsightModel
            }
        }
        duplicateInsightSuccess: (item: QueryBasedInsightModel) => {
            type: 'duplicate insight success (models.insightsModel)'
            payload: {
                item: QueryBasedInsightModel
            }
        }
        insightsAddedToDashboard: ({ dashboardId, insightIds }: { dashboardId: number; insightIds: number[] }) => {
            type: 'insights added to dashboard (models.insightsModel)'
            payload: {
                dashboardId: number
                insightIds: number[]
            }
        }
    }
    actionKeys: {
        'rename insight (models.insightsModel)': 'renameInsight'
        'rename insight success (models.insightsModel)': 'renameInsightSuccess'
        'duplicate insight (models.insightsModel)': 'duplicateInsight'
        'duplicate insight success (models.insightsModel)': 'duplicateInsightSuccess'
        'insights added to dashboard (models.insightsModel)': 'insightsAddedToDashboard'
    }
    actionTypes: {
        renameInsight: 'rename insight (models.insightsModel)'
        renameInsightSuccess: 'rename insight success (models.insightsModel)'
        duplicateInsight: 'duplicate insight (models.insightsModel)'
        duplicateInsightSuccess: 'duplicate insight success (models.insightsModel)'
        insightsAddedToDashboard: 'insights added to dashboard (models.insightsModel)'
    }
    actions: {
        renameInsight: (item: QueryBasedInsightModel) => void
        renameInsightSuccess: (item: QueryBasedInsightModel) => void
        duplicateInsight: (item: QueryBasedInsightModel) => void
        duplicateInsightSuccess: (item: QueryBasedInsightModel) => void
        insightsAddedToDashboard: ({ dashboardId, insightIds }: { dashboardId: number; insightIds: number[] }) => void
    }
    asyncActions: {
        renameInsight: (item: QueryBasedInsightModel) => Promise<any>
        renameInsightSuccess: (item: QueryBasedInsightModel) => Promise<any>
        duplicateInsight: (item: QueryBasedInsightModel) => Promise<any>
        duplicateInsightSuccess: (item: QueryBasedInsightModel) => Promise<any>
        insightsAddedToDashboard: ({
            dashboardId,
            insightIds,
        }: {
            dashboardId: number
            insightIds: number[]
        }) => Promise<any>
    }
    defaults: {}
    events: {}
    key: undefined
    listeners: {
        renameInsight: ((
            action: {
                type: 'rename insight (models.insightsModel)'
                payload: {
                    item: QueryBasedInsightModel
                }
            },
            previousState: any
        ) => void | Promise<void>)[]
        duplicateInsight: ((
            action: {
                type: 'duplicate insight (models.insightsModel)'
                payload: {
                    item: QueryBasedInsightModel
                }
            },
            previousState: any
        ) => void | Promise<void>)[]
    }
    path: ['models', 'insightsModel']
    pathString: 'models.insightsModel'
    props: Record<string, unknown>
    reducer: (state: any, action: any, fullState: any) => {}
    reducers: {}
    selector: (state: any) => {}
    selectors: {}
    sharedListeners: {}
    values: {}
    _isKea: true
    _isKeaWithKey: false
}